requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
Pillow==10.2.0
textblob==0.17.1
requests==2.31.0
beautifulsoup4==4.12.3
openai==1.52.0
python-dotenv==1.0.1
google-generativeai==0.8.3
ollama
